perf: profile page
