+++ package.json
@@ -29,6 +29,7 @@
     "zustand": "^4.5.5"
   },
   "devDependencies": {
+    "@tanstack/config": "^0.13.1",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
