+++ src/types/next-auth.d.ts
@@ -3,8 +3,10 @@ import 'next-auth';
 declare module 'next-auth' {
   interface Session {
     user: {
-      id: number;
+      employeeId: number;
       name: string;
+      surname1: string;
+      surname2: string;
       email: string;
       access_token: string;
     };
